分離
