+++ crates/voicevox_core_c_api/Cargo.toml
@@ -44,3 +44,4 @@ regex.workspace = true
 serde.workspace = true
 toml = "0.7.2"
 typetag = "0.2.5"
+inventory = "0.3.4"
